updates on script

+++ bin/signaged.rb
@@ -32,7 +32,7 @@ $content_dir = "#{base_dir}/../downloads"
 puts $PROGRAM_NAME + ": Using content dir " + $content_dir
 
 # A loop that downloads and show the files
-# last_schedule_id = -1
+last_schedule_id = -1
 player_pid = -1
 
 while true
@@ -42,31 +42,31 @@ while true
     current_schedule = synchronizer.sync
   end
 
-  # if current_schedule['id'] != last_schedule_id
-  player_pid = begin
-                 File.read([$content_dir, 'signaged_player.pid'].join('/')).to_i
-               rescue
-                 -1
-               end
+  if current_schedule['id'] != last_schedule_id
+    player_pid = begin
+                   File.read([$content_dir, 'signaged_player.pid'].join('/')).to_i
+                 rescue
+                   -1
+                 end
 
-  puts "PID on File: '#{player_pid}'"
+    puts "PID on File: '#{player_pid}'"
 
-  if player_pid > 0
-    # ask the player to stop if it's running
-    begin
-      Process.kill("TERM", player_pid)
-      Process.wait
-    rescue
-      puts "PID #{player_pid} not found."
+    if player_pid > 0
+      # ask the player to stop if it's running
+      begin
+        Process.kill("TERM", player_pid)
+        Process.wait
+      rescue
+        puts "PID #{player_pid} not found."
+      end
     end
+
+    # spawn the player
+    serialized_items = Shellwords.escape JSON.generate(current_schedule['items'])
+    spawn("#{base_dir}/signaged_player.rb #{serialized_items}")
   end
 
-  # spawn the player
-  serialized_items = Shellwords.escape JSON.generate(current_schedule['items'])
-  spawn("#{base_dir}/signaged_player.rb #{serialized_items}")
-  # end
-
-  # last_schedule_id = current_schedule['id']
+  last_schedule_id = current_schedule['id']
 
   current_schedule = synchronizer.sync
